Also export a list of all the files affected by a pull request
To maintain backwards compatibility, a new variable containing all the changed files is introduced.

Given that, the variable `all_changed_files` was already taken and we do not want to break backwards compatibility. And also, the prefix `all` is pretty inclusive already, a rather difficult decision needed to be taken.

So, I didn't take any. This is a placeholder name for the new super all inclusive variable name that contains all the files, absolutely all the files, no file left behind, cross my heart and hope to die.

We are however open to suggestions!